Update README for case sensitivity bug
Many guides around the internet, including in this repositories own issue history, suggest using `gcloud config get-value account` to get the email address. The --user option is case sensitive with respect to the email address provided. This will cause the `sc install` to fail with a verbose permissions error that makes it look like the clusterrolebinding was not created. This message will give guidance that users should check the casing of the email address provided.